small change in parsing

+++ TTplugin-PhyIO/com.testingtech.ttworkbench.phyio.arduino/PhyIOCarSim/led/flexibelLEDFunction/flexibelLEDFunction.ino
@@ -183,7 +183,7 @@ void LEDFunction( int id, int command) {
       LEDFunctionSet(id, XSERIAL.parseInt());
       break;
     case BLINK:
-      LEDFunctionBlink(id, (1000/(XSERIAL.parseInt()*2)));
+      LEDFunctionBlink(id, (1000/(XSERIAL.parseFloat()*2)));
       break;
     case STOP:
       LEDFunctionSet(id, 0);
